fix: fix clang-tidy warnings

+++ src/common/egraph_config.cpp
@@ -14,9 +14,9 @@ void initialize_config_for_expression(EGraphConfig &config, const Expression &ex
 
     // Scale iteration depth based on AST complexity
     config.rewrite.max_iterations =
-        std::max(config.rewrite.max_iterations, std::max(static_cast<size_t>(10), static_cast<size_t>(depth * 1.6)));
-    config.pruner.rewrite_steps_per_iteration = std::max(static_cast<size_t>(10), static_cast<size_t>(depth * 2));
-    config.pruner.prune_samples_per_iteration = std::max(static_cast<size_t>(10), static_cast<size_t>(depth * 12));
+        std::max(config.rewrite.max_iterations, std::max(10, static_cast<int>(static_cast<double>(depth) * 1.6)));
+    config.pruner.rewrite_steps_per_iteration = std::max(10, static_cast<int>(depth * 2));
+    config.pruner.prune_samples_per_iteration = std::max(10, static_cast<int>(depth * 12));
     config.pruner.dag_visit_limit = std::max(static_cast<size_t>(10000), static_cast<size_t>(depth * 1500));
     config.pruner.dag_sample_ratio = 0.05;
 }
